+++ lib/papa.rb
@@ -1,11 +1,13 @@
 require 'thor'
 require 'papa/thor'
+require 'papa/string'
 require 'papa/version'
 require 'papa/cli'
 require 'papa/command'
 require 'papa/command_queue'
 require 'papa/git'
 require 'papa/output'
+require 'papa/helpers'
 
 module Papa
 end