+++ lib/papa/cli/sandbox.rb
@@ -2,11 +2,16 @@ module Papa
   class Sandbox < Thor
     desc 'generate', 'Generate a sandbox environment'
     option :override_origin, aliases: '-o'
+    option :override_path_prefix, aliases: '-p'
     def generate
-      override_origin = options[:override_origin]
-
       require 'papa/sandbox/generate'
-      Sandbox::Generate.new(override_origin: override_origin).run
+      Sandbox::Generate.new(options).run
+    end
+
+    desc 'clean', 'Clean up sandbox directories in /tmp'
+    def clean
+      require 'papa/sandbox/clean'
+      Sandbox::Clean.new.run
     end
   end
 end