+++ lib/papa/common/start.rb
@@ -3,13 +3,14 @@ module Papa
     def run
       @build_branch ||= "#{@build_type}/#{@version}"
 
-      queue = CommandQueue.new
-      queue.add Git.fetch(remote: 'origin')
-      queue.add Git.checkout(branch_name: @base_branch)
-      queue.add Git.branch(branch_name: @build_branch)
-      queue.add Git.checkout(branch_name: @build_branch)
-      queue.add Git.push(remote: 'origin', branch_name: @build_branch)
-      if queue.run
+      @queue = CommandQueue.new
+      @queue.add Git.fetch(remote: 'origin')
+      @queue.add Git.checkout(branch_name: @base_branch)
+      @queue.add Git.branch(branch_name: @build_branch)
+      @queue.add Git.checkout(branch_name: @build_branch)
+      @queue.add Git.push(remote: 'origin', branch_name: @build_branch)
+
+      if @queue.run
         success_message
       else
         failure_message
@@ -20,11 +21,11 @@ module Papa
     private
 
     def success_message
-      Output.stdout "Successfully started new #{@build_type} branch #{@build_branch}"
+      Output.success "Successfully started new #{@build_type} branch #{@build_branch}"
     end
 
     def failure_message
-      Output.stderr "There was a problem starting #{@build_type} branch: #{@build_branch}"
+      Output.failure "There was a problem starting #{@build_type} branch: #{@build_branch}"
     end
   end
 end